refactor: made sorting survey more readable, reused functionality accross files and isolated tasks
