+++ lib/core/routing/app_routes.dart
@@ -1,7 +1,7 @@
-import 'package:edconnect_admin/presentation/pages/sorting_module_pages/create_sorting_survey_page.dart';
+import 'package:edconnect_admin/presentation/pages/sorting_module/pages/create/create_sorting_survey_page.dart';
 import 'package:flutter/material.dart';
-import 'package:edconnect_admin/presentation/pages/sorting_module_pages/sorting_surveys_page.dart';
-import 'package:edconnect_admin/presentation/pages/sorting_module_pages/sorting_survey_details.dart';
+import 'package:edconnect_admin/presentation/pages/sorting_module/pages/list/sorting_surveys_page.dart';
+import 'package:edconnect_admin/presentation/pages/sorting_module/pages/details/sorting_survey_details.dart';
 
 class AppRoutes {
   // Prevent instantiation
