+++ lib/data/repositories/navigation_repository_impl.dart
@@ -4,7 +4,7 @@ import '../../domain/entities/navigation_item.dart';
 import '../../domain/entities/permissions.dart';
 import '../../presentation/pages/auth_pages/access_denied_page.dart';
 import '../../presentation/pages/sample_page.dart';
-import 'package:edconnect_admin/presentation/pages/sorting_module_pages/sorting_surveys_page.dart';
+import 'package:edconnect_admin/presentation/pages/sorting_module/pages/list/sorting_surveys_page.dart';
 
 class NavigationRepositoryImpl implements NavigationRepository {
   // Cache navigation items by ID for faster lookup
