+++ lib/presentation/pages/sorting_module/components/build_info_row.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:edconnect_admin/core/design_system/foundations.dart';
+import 'package:edconnect_admin/presentation/providers/theme_provider.dart';
+
+class InfoRow extends ConsumerWidget {
+  final String label;
+  final String value;
+  final Widget? trailing;
+  final VoidCallback? onTap;
+
+  const InfoRow({
+    required this.label,
+    required this.value,
+    this.trailing,
+    this.onTap,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final theme = ref.watch(appThemeProvider);
+    final isDarkMode = theme.isDarkMode;
+
+    Widget content = Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(
+          width: 160,
+          child: Text(
+            label,
+            style: TextStyle(
+              fontSize: Foundations.typography.base,
+              fontWeight: Foundations.typography.medium,
+              color: isDarkMode
+                  ? Foundations.darkColors.textSecondary
+                  : Foundations.colors.textSecondary,
+            ),
+          ),
+        ),
+        Expanded(
+          child: Row(
+            children: [
+              Expanded(
+                child: Text(
+                  value,
+                  style: TextStyle(
+                    fontSize: Foundations.typography.base,
+                    color: isDarkMode
+                        ? Foundations.darkColors.textPrimary
+                        : Foundations.colors.textPrimary,
+                  ),
+                ),
+              ),
+              if (trailing != null) trailing!,
+              if (onTap != null)
+                Icon(
+                  Icons.chevron_right,
+                  color: isDarkMode
+                      ? Foundations.darkColors.textMuted
+                      : Foundations.colors.textMuted,
+                ),
+            ],
+          ),
+        ),
+      ],
+    );
+
+    if (onTap != null) {
+      return InkWell(
+        onTap: onTap,
+        borderRadius: Foundations.borders.md,
+        child: Padding(
+          padding: EdgeInsets.symmetric(vertical: Foundations.spacing.xs),
+          child: content,
+        ),
+      );
+    }
+
+    return content;
+  }
+}
