+++ lib/presentation/pages/sorting_module/tabs/responses/components/stats/stat_card.dart
@@ -1,0 +1,73 @@
+import 'package:edconnect_admin/core/design_system/foundations.dart';
+import 'package:edconnect_admin/presentation/providers/theme_provider.dart';
+import 'package:edconnect_admin/presentation/widgets/common/cards/base_card.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class StatCard extends ConsumerWidget {
+  final String label;
+  final String value;
+  final IconData icon;
+  final double width;
+
+  const StatCard({
+    super.key,
+    required this.label,
+    required this.value,
+    required this.icon,
+    required this.width,
+  });
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final theme = ref.watch(appThemeProvider);
+    return SizedBox(
+      width: width,
+      child: BaseCard(
+        variant: CardVariant.outlined,
+        margin: EdgeInsets.zero,
+        padding: EdgeInsets.all(Foundations.spacing.md),
+        child: Row(
+          children: [
+            Icon(
+              icon,
+              size: 24,
+              color: theme.isDarkMode
+                  ? Foundations.darkColors.textMuted
+                  : Foundations.colors.textMuted,
+            ),
+            SizedBox(width: Foundations.spacing.md),
+            Expanded(
+              child: Column(
+                mainAxisSize: MainAxisSize.min, // Use minimum height
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    label,
+                    style: TextStyle(
+                      fontSize: Foundations.typography.sm,
+                      color: theme.isDarkMode
+                          ? Foundations.darkColors.textMuted
+                          : Foundations.colors.textMuted,
+                    ),
+                  ),
+                  SizedBox(height: Foundations.spacing.xs),
+                  Text(
+                    value,
+                    style: TextStyle(
+                      fontSize: Foundations.typography.lg,
+                      fontWeight: Foundations.typography.semibold,
+                      color: theme.isDarkMode
+                          ? Foundations.darkColors.textPrimary
+                          : Foundations.colors.textPrimary,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
